+++ Examples/MultiApple/iOS/main.m
@@ -11,7 +11,7 @@
 int main(int argc, char *argv[]) {
     
     NSAutoreleasePool * pool = [[NSAutoreleasePool alloc] init];
-    int retVal = UIApplicationMain(argc, argv, nil, nil);
+    NSInteger retVal = UIApplicationMain(argc, argv, nil, nil);
     [pool release];
     return retVal;
 }
