+++ UIKit/Classes/UIScrollView.m
@@ -42,6 +42,7 @@
 #import "UIScrollWheelGestureRecognizer.h"
 #import "UIGeometry.h"
 #import <QuartzCore/QuartzCore.h>
+#include <tgmath.h>
 
 static const NSTimeInterval UIScrollViewAnimationDuration = 0.33;
 static const NSTimeInterval UIScrollViewQuickAnimationDuration = 0.22;
@@ -421,8 +422,8 @@ static NSString* const kUIScrollIndicatorInsetsKey = @"UIScrollIndicatorInsets";
         [self _setScrollAnimation:animation];
         [animation release];
     } else {
-        _contentOffset.x = roundf(theOffset.x);
-        _contentOffset.y = roundf(theOffset.y);
+        _contentOffset.x = round(theOffset.x);
+        _contentOffset.y = round(theOffset.y);
 
         [self _updateContentLayout];
 
@@ -492,9 +493,9 @@ static NSString* const kUIScrollIndicatorInsetsKey = @"UIScrollIndicatorInsets";
 - (UIScrollViewAnimation *)_pageSnapAnimation
 {
     const CGSize pageSize = self.bounds.size;
-    const CGSize numberOfWholePages = CGSizeMake(floorf(_contentSize.width/pageSize.width), floorf(_contentSize.height/pageSize.height));
+    const CGSize numberOfWholePages = CGSizeMake(floor(_contentSize.width/pageSize.width), floor(_contentSize.height/pageSize.height));
     const CGSize currentRawPage = CGSizeMake(_contentOffset.x/pageSize.width, _contentOffset.y/pageSize.height);
-    const CGSize currentPage = CGSizeMake(floorf(currentRawPage.width), floorf(currentRawPage.height));
+    const CGSize currentPage = CGSizeMake(floor(currentRawPage.width), floor(currentRawPage.height));
     const CGSize currentPagePercentage = CGSizeMake(1-(currentRawPage.width-currentPage.width), 1-(currentRawPage.height-currentPage.height));
     
     CGPoint finalContentOffset = CGPointZero;
